ui: Defer invoking `Track.onCreate()` until the track is in view
Currently we call onCreate() the first time a track appears on the timeline, regardless of whether or not the track is scrolled into view.

However, a lot of tracks defer expensive queries until the first onUpdate() call, which is only called when the track is actually scrolled into view.

This CL changes the track FSM behavior so that onCreate() is only called the first time a track is scrolled into view - i.e. right before the first call to onUpdate(). This means tracks can make better use of this lifecycle hook, and don't have to manage this state internally.

Change-Id: Icc6b59dbf3730cc9cd6f15822ca263fbf645f76b