ui: overhaul track decider queries to make things faster
This CL overhauls a bunch of queries in track_decider to improve
load performance.

Specifically, the main thing we are doing is extracting a bunch of O(event)
queries into the standard library and compute a summary over each of the
event tables exactly once. This significantly reduces the amount of
redundant computation that needs to be performed.

Change-Id: I49c07f681c7128cf37852ad66a8a2a8e2e17b2b5