ui: Add virtual scroll widget, decouple query vs DOM skirting
In the current implementation of virtual scrolling used in ftrace & android logs tabs, we query a subset of rows from the SQL table, enough to cover the the viewport + some excess above and below the viewport. This excess is known as a "skirt".

The size of the skirt needs to be carefully tuned. Too small and the user sees rows popping in as they're scrolling, if queries or event handlers take a long time. Too large, and we end up drawing too many DOM elements that the user isn't going to see.

Because it's more efficient to query more rows in fewer queries, we currently make this skirt large. The actual implementation uses paging technique with pages sized at 250 rows, so we usually end up loading and rendering ~3 pages = 750 rows, which is a LOT more than is required to fill in the visible window. Rendering this many excess rows can destroy performance.

We have a few options:
- A: Reduce the size of the skirt.
- B: Add an additional smaller skirt to render only a subset of queried rows to the DOM.

Option A results in more flickering and pop-in when scrolling.

Option B offers the best of both worlds. Thus, this CL implements option B.

I have also created a new widget `VirtualTable` which manages the logic behind the two different skirts as well as applying consistent styling and features e.g. sticky headers.

Bug: 335451611
Change-Id: I4cb0a236c8e8afc9d3757ebadfef0a7d5787ec3a