Trace Redaction - Assign empty comm for cleared values
Before, we removed the next_comm and prev_comm values by not copying
the fields over. This caused issues in the perfetto ui as it would
use "a value" causing names to appear from "somewhere".

Now, we replace strings with an empty string, fixing the issue. We've
seen this issue elsewhere and should be applied to future
modifications.

Bug: 336807771
Change-Id: Ia7e6cab076e0ec0de7f2cdedb1903f1da2040172