Trace Redaction - Replace comm instead of drop comm in proc free
With sched switch and waking events, we've seen issues dropping comm
fields. This issue goes away when they are replaced with an empty
string.

Because proc free events are few, it is hard to find issues in the
Perfetto UI. To be proactive, this updates the proc free primitive to
use the empty string approach.

Change-Id: I29bcc6efdb6081648d9144606af5b540b9a5dc26