Remove PacketSequenceState
Rename PacketSequenceState to PacketSequenceStateBuilder and hide it
in the ProtoTraceReader. Only this class should use it.

Make it more clear what state is keep around and what state is reset
for the various sequence events (packet loss, incremental state
cleared, etc).

This also gives us tow different API surfaces one for ProtoTraceReader
instances to update the incremental state as packets are porcessed in
sequence order: PacketSeqeunceStateBuilder, and another "read only" API
to be used by tokenizers and parsers of concrete trace packets.

Change-Id: I64981bddd2058e48b6e4d9bd3095a4c0bb41063e