Trace Redaction - Remap pids to synth pids
Whenever a pid maps to a process/thread outside of the target package,
replace it with the pid reserved for the cpu hosting the thread.

This will cause threads in the process tree to have no sched waking or
sched switch events. To remove those entries, an additional primitive
will be needed.

Because synth threads don't appear in the process tree, the synth
threads won't appear in the same process, so each will be treated as
a main thread.

Bug: 336807771
Change-Id: Idd87f7773ebc94569e086b6c33db07994099bf4e