Trace Redaction - Remove Timeline Optimizations
There were some timeline optimizations that would limit the depth
of each query. While this works when queries were to a single uid,
it does not work when queries need to work for multiple uids or pids.

This becomes relevant when addition (system) processes need to be
retained (e.g. Surface Flinger).

Bug: 333740194
Change-Id: I6b79dace64f57932d9a8bd952b00256a81dedd0d