Trace Redaction - Simplify and expand usage of timeline event
A timeline event had a lot of boiler plate code and was often wrapped
in std::optional. The timeline type also had an invalid option, but
it was never used.

This CL strips the event type to be a simple struct and uses the event
invalid type instead of std::optional.

The timelines interface does not change, but uid values change because
invalid pid and uid values change from 0 (a valid value) to max
int32_t and max uint64_t. This exposed errors in the tests.

Bug: 333740194
Change-Id: Iad1d51bea7356b36858cb841389410c61d419db6