Trace Redaction - Expose and test FindPreviousEvent
The timeline is a collection of events. The primary operation it has
is "find previous event", a function that finds the most recent
event (relevant to some time) for a process and/or thread.

This function is used to find a process's / thread's package. The
only way to test FindPreviousEvent() was through the search function.

To improve testing, this exposes FindPreviousEvent() so it can be
tested and used directly.

Change-Id: I7991fb3c7e7f8f6cf13c0fd863f105c20a38226f