Build winscope extensions proto only as "source_set" and "zero"
Specifically, the "cpp" build, which produces ".gen.cc" and ".gen.h"
is not supported for those files, because (among other things) they
reference "google/protobuf/descriptor.proto", which is not correctly
wired up for gen files.

The existence of the "cpp" targets was not a problem for most perfetto
builds because, if a target is not used, it's not build. The chromium
CI, on the other hand, tries to build every target, even if it's not
referenced.

Change-Id: Ida476f01b499acf65a0382242a733c4dca82f3c7