Create TimelineSync plugin
This plugin allows to syncronize the timeline of several open windows. Even if there are multiple different traces, the initial viewport (when the command is enabled) is considered the starting point, and any movement/zoom operation is applied to each window.

This is useful to compare a good and bad trace side to side.

Bug: 337016334
Bug: 338585751
Test: Run plugin with 2 windows opened
Change-Id: Ie16479cf475b9997f7f940186987f33268a47603