ui: avoid multi-reload on live dev server
Today when saving a .ts file we often get multiple reloads of the UI.
This is because the current rules trigger a livereload for every
file change in dist/. What happens when you edit a .ts file is that:
- frontend_bundle.js or engine_bundle.js gets updated.
- The source map (e.g., frontend_bundle.js.map) gets updated.
- The manifest.json (for serviceworker) get updated.

In 99% of cases developers care only about the first one.
Filtering out notification and propagating only js|css|wasm|html changes.

Change-Id: Iad5720b19004042d791ecdd0fe8ecd9c8155642e