ui: Fix crash when dragging a track out of a group
Dragging a track out of a group effectively makes a copy of the track,
with the same key. While there's nothing technically wrong with having
multiple tracks with the same key* this key was also being used to
define the panel ID, which much be unique.

Panels are an intermediate object between tracks and Mithril
components, representing a generic horizontal slice of screen real
estate on the timeline.

In contrast to tracks, panels must have unique IDs because these IDs
are stored in DOM elements and used to resolve back to panel objects.

  Track[key] -> Panel[id] -> Mithril -> DOM[data-panel-id]
                     ^                        /
                      `-----------------------

* Having multiple tracks with the same key results in odd area
  selection behaviour, but this is no worse than it was before.
  We can tackle this in another CL.

The root cause is that the panel/track/group ID/key convention is
confusing:
- Mix of "id" and "key" terminology in tracks and groups.
- Panels have keys and trackKeys and trackGroupIds which have differing
  uniqueness requirements, used for a mix of different things.

This CL:
- Renames all instances of "trackGroupId" to simply "groupKey". This is
  for consistency with "trackKey".
- Don't bother setting keys on mithril components for each panel as
  these were being generated from array indices and thus served no
  purpose, as mithril keys are only useful if they're consistent
  between render cycles.
- Remove key (i.e. panelId) from panels - leave trackKey and groupKey.
  Instead, generate panel ids on the fly from array indices for panel
  ids - these need to be unique but don't need to be consistent over
  render cycles.

Bug: https://github.com/google/perfetto/issues/791
Bug: 341429501
Change-Id: I51ae9b9606e91fdb1e64b8807bc60299126145aa