tp: improve handling of !usable constraints in BestIndex
This CL changes how !usable constraints are handling to respect the
SQLite documentation better: the documentation has changed since this
behaviour was originally implemented.

Specifically, we return errors more eagerly if !usable constraint is
present for a required argument. This allows SQLite to know that we
need the column to be present. We also return proper error messages
if a required constraint is completely missing.

See
https://www.sqlite.org/vtab.html#enforcing_required_parameters_on_table_valued_functions
for the documentation on this.

Change-Id: If927e0a890e62cde7dd9c2114144028f25081589