Trace Redaction - Create base for sched redaction
The are multiple places where scheduling information exists and the
comm and pid values need to be modified.

There are two levels to this, there is the redaction logic. This is
the logic that handles "given pid P and comm C, what should it change
to". This can be called the "modification". The other part controls
when/where these modifications are invoked (the harness).

In this change, the old primitive is replaced with a harness and
a single modification (clear comms). The number of modifications and
the scope of the harness will grow as new modifications are created
(e.g. thread merging) and new locations are uncovered.

Bug: 336807771
Bug: 339504266
Change-Id: Ibaee826a955576c5818d1ec99bc7c230caf47731