ui: improve TimelineSync plugin
Rather than requiring to run the command on all tabs,
show a list of tabs and do everything from one trace.
This adds an advertise method on the broadcast channel
to allow each instance to discover the other ones.
The final result is that a single invocation of the
command is enough to keep N tabs in sync, without
requiring the user to manually run the command on
each tab.
Also support concurrent sessions between disjoint
sets of tab groups.

Screenshot: https://screenshot.googleplex.com/BNfmVvNvFJkuAmx

Change-Id: I00fff31de425b4ca2aad563696d3014286c097b0