Trace Redaction - Redact CompactSched Switch Events
CompactSched switch events are expressed using assosicative arrays
along side an intern string table (shared with waking events).

To redact these events, the pid and comm arrays need to be "zipped"
together and passed through a transform. Compared to ftrace switch
events, the transform is non-trival due to the intern string table.

To transform the comm value, the old comm value needs to be
retrived from the intern table, optionally transformed, and then
conditionally appended to the internt able (if it was not in the
intern table).

Because the compact sched switch events are in a ftrace bundle, all
events occur on the cpu listed on the bundle.

Bug: 339504266
Change-Id: Ia46746dbc19c3ddda74ce41fd5a5de6d67f73871