Trace Redaction - Re-implement proc free redaction
The primitve that redacted proc free events was vague, and hid what it
was doing. This renames the primitive to be clearer about what its
doing, removing the comm value.

The implementation was more complicated than it needed to be. There
are only three fields in a free event, but the implementation was
acting as if there were many more. This changing the implementation
to explicitly reference each value. By doing this, the integrity of
the event can be verified.

Bug: 338425440
Change-Id: I94dce12df1153d90ba92bf17c956367697410e4c