Support multiple trace ingestion via ZIP files
Add support to parse ZIP files. TP will detect the type of trace file
contained in the ZIP file and parse it accordingly.

This cl also adds support to parse multiple trace files if multiple
files are contained in the ZIP. There is a well defined order in which
traces are parsed, first proto traces (because these might contain clock
sync events) and then the rest.

There is one exception (which we might remove in the future), proto
traces with ModuleSymbol packets are parsed last (this is needed so that
TP correctly applies the symbols)

Bug: b/334978369
Change-Id: Iaf5d2eca968c6ca4c4074ba30b106467cde479ce