tp: fix tracking when process reuse happens before ftrace starts
This CL fixes a subtle issue with process tracking when a new process is
spawned *before* we have ftrace events telling us exactly what processes
died and restarted. In this case, we were not correctly creating new
processes based on the parent pid changing. There might still be an edge
case here with processes getting reparented to init but that seems a
very rare edge case so we don't tackle it until we see an example of
that.

Change-Id: Ib68ef1159d852cd33f2761329a349a6045586983
Bug: 339301226