perfetto: add a special event for traces started using clone
This CL adds a special event for traces start using --clone in
perfetto_cmd. Instead of using TraceBegin/BackgroundTraceBegin
we instead have special CloneTrace and CloneTriggerTrace events.

Also for CloneTrigger traces, plumb through the trigger name
from the service back to perfetto_cmd so it can be emitted in the
atom.

Change-Id: I113f5f8e8c008ac621582ad809a174bc80394bee
Bug: 289200715