tp: refactor CPU-indexed tables with the 'ucpu' column
This CL introduces the 'ucpu' column, which is the id of the 'cpu' table
for uniquely identifying CPUs in the multiple-machine tracing context.
The following tables are refactored: cpu, cpu_freq, sched_slice, raw,
ftrace_event, experimental_sched_upid, thread_state. The cpu table is
inserted not only on cpu_info trace packets, but also when CPU-indexed
trace data are injested.

This allows for more efficient query for listing CPU scheduling tracks
in the UI:

* Original: 'select distinct cpu from sched', which scans the sched
table only for distinct cpu indexes
* Update: 'select cpu from cpu', which scans a much smaller cpu table.

Bug: 284258446
Test: diff_tests

Change-Id: I895e1cb4b90c92277bb42b78c01950883641a8c4