Trace Redaction - Override trusted uids
In production, there should only be two trusted uids (1000 and 9999),
but this makes it hard to test redaction because most traces will have
different uids.

To make it easy to programmatically override it, this moves the strict
"is A or B" to use an allowlist. This allowlist can be changed,
changing which uids are trusted. To be safe, the default is set to
1000 and 9999.

Change-Id: I5a2feb2d1e8451892ed9b21ff16cfc70cdb3f1ea