ui: merge FlamegraphController into DetailsPanel
This CL merges all the logic in flamegraph controller into the details
panel and generally makes things a lot eaiser to follow. This is a pure
refactoring CL: there should be no user-visible changes as a result of
this CL.

Future CLs will further improve the logic here, especially on the
queries. This requires trace processor changes.

Change-Id: Ice9800e3a1d64a3b3822a20e46fbcefb710df736