Trace Redaction - Move 'redact print' to 'redact process events'
For thread merging to work, the ftrace event pid need to be
redacted. This will cause issues around order of operations because
once it is changed, later transform will have the merged pid.

By moving more into this one primitive, it should be easier to
reduce risk, even if it is breaking the "many small operations"
pattern.

Change-Id: Ied18fb3538b384b1c9e205412258ed43ece01c15