Trace Redaction - Use strategy pattern for ftrace events
Following changes else where in the code base, this changes the
top-level ftrace event primitive to use the strategy pattern
to filter and write ftrace events to the new packet.

The "write" function will be used to change the ftrace event pid
during thread merging.

The old remapping primitives had to be removed from the build files
because this change conflicted with them (class designs). Because
these files/classes were going to be revisited, this cost was
accessible.

The ReturnErrorForEmptyAllowList test was removed because it is no
longer a valid test. The allowlist test is done in one of the filters.
Filters cannot return errors. To support that, it would introduce
additional overhead for very little value. If the need to verify
allowlists is so great, build primitive can be used.

Bug: 336807771
Change-Id: I15e4ab548dd985304daec8c1febcf241203980f2