Add support for processing interned protolog ViewerConfigs
We want to support protologging without necessarily processing the source files to generate the Viewer Configs that we dump at then end of a trace. So instead we need to dump the ViewerConfig to decode the message at runtime in the incremental state. In a way the ViewerConfig proto serves as a way to intern protolog messages to save memory.

Test: tools/ninja -C out/linux_clang_release && tools/diff_test_trace_processor.py ./out/linux_clang_release/trace_processor_shell --name-filter "ProtoLog"
Change-Id: I5a5a303ccd5e8ae14435ce031781d4d17666bf39