tracing_service_impl: Reply to clone consumer if src session ends
Today, if a consumer issues a clone request and the original tracing
session goes away, the original session does not get a reply.

This commit fixes the problem.

The problem was there since the original clone implementation.
perfetto_cmd has a 30s timeout on the for a clone operation, so this was
not a huge problem.

Change-Id: Ib741033fc1db5525c6c96812ecdbe88b7e523d5e