Trace Redaction - Make Timeline Range Inclusive
Before, the timeline would treat open events as inclusive bounds and
would treat close events as exclusive. This creates problems.

Process free events are the only way a thread can end. However, if
close events are exclusive, a thread that should end would not be
connected to the target package.

The previous implementation would look for the latest, previous event.
When given a close event, there was no way to test the ppid or the
uid.

To fulfill the requirements, the internal implementation was changed
to look up the latest, previous open and the latest, previous close
event. Using these two events all the necessary information is
available.

Change-Id: Icaa798a262d6b73029f3e7af3cdd8cf9817999a5