Trace Redaction - Move suspend/resume into process events
To make it easier for suspend/resume events to be filtered, this CL
exposes the events with the 'redact process events' class. This
class is already handling other events, like new task and process
free, so it seems like a better home for it.

By doing this, it makes merging threads easier by limiting the what
need to be configured correctly.

Change-Id: If5a7cebf4dce1cb227e6171c8b32871de82f2ded