Move Nettrace events to slice subtable
This change adds a slice subtable for nettrace events. This allows
attaching all of the attributes directly in the table which can be much
faster to query. The args are left as-is for now since they are how the
details are presented to the user.

This does add slight overhead (~2%) to the loading phase, but can speed
up a number of queries by more than 2x.

Bug: 332311721
Bug: 341994378
Change-Id: Ib4f1d146a65200ca6aba1d08ce8706b5e442c245