Fix the network activity query correctness
This works by finding gaps in events. The assumption was that the last
ts+dir was the largest one. It is possibel for events to overlap in such
a way that isn't true. For example:

```
ts   dur
1    5
3    1
16   2
```

The old method should show t=16 having a 16-(3+1) = 12 gap, but the
correct gap is 16-(1+5) = 10 gap. This could change group boundaries,
creating overlaps or wrong totals.

Change-Id: I0c71e4f2b1551687c48c6aadba226cd51d4af642