shared_lib: Fix PERFETTO_TE_SCOPED macro captures
The implementation had a bug which made it basically unusable.

```
uint64_t v = 5;
PERFETTO_TE_SCOPED(cat, PERFETTO_SLICE("slice"),
                   PERFETTO_TE_ARG_UINT64("arg", v));
```

Would expand to

```
uint64_t v = 5;
class PerfettoTeEvent {
  struct Internal {
    Internal() {
      PERFETTO_TE(cat, PERFETTO_SLICE("slice"), PERFETTO_TE_ARG_UINT64("arg", v));
    }
    //...
```

But `v` would not be accessible inside the class and the struct.

This commit switches to a lambda that captures everything to solve the
problem (the C++ macros uses a similar solution).

Change-Id: I79023c5718679495b7fa56bf658b649d9f4b1da3