Trace Redaction - Redact Sched Blocked Reason
Just like waking events, sched bloced reason events contain an outer
and inner pid. The outer pid represents the thread blocking the
inner thread.

Just like waking events, it is okay for the outer pid to be replaced
with a synth thread, but the inner thread cannot be replaced with an
inner thread. To address this issue, sched blocked reason events are
only retained if the inner pid belongs to the target package.

Bug: 344064721
Change-Id: Ife80781ff10e44cac224fc6d83f29cf78fbac6a3