ui: Migrate counter selection
This patch migrates the counter selection from the old legacy selection
mechanism to the new single event selection mechanism.

In this new world, selections are tightly coupled to tracks, rather
than being distinguished by selection kind. This means that the track
implementation teaches the UI how to render the details panel and how
to find the bounds of the selected entity through additional arguments
on the TrackDescriptor passed to ctx.registerTrack().

This also means that `SelectionController` no longer needs to load
counter details, or publish them to globals, so this was removed.

Counter tracks are a simple one to get started with as there aren't
that many of them, and their details panels are relatively simple
compared to slice tracks.

Testing:
- Select a counter on a counter track, check that the correct details
  panel appears.
- Hit 'F' while a counter sample is selected, check that the slice is
  centred, and hitting it again zooms into the correct bounds.
- Hit 'M' while a counter sample is selected, check that the resulting
  span note has the correct bounds.

Change-Id: I9f4d2a559852d413a0f6d9017a0551151e8b8250