ui: Register a separate track object for every debug track created
Currently we register a single debug track and instantiate it
multiple times, providing the debug track's state via the config
object within the immutable store.

This patch makes it so we register a new track object with a distinct
URI each time we add a new debug track.

This ultimately allows us to move forward with the new selection types
which hook details panels onto track descriptors.

It also gets us closer to the ideal where we no longer rely on track
state from the immutable store to customize the track details.

Testing:
- Create slice and counter debug tracks using the "AddDebugTrackMenu"
- Check that debug slices can be selected and the details panels still
  show up
- Check critical path commands and buttons still produce tracks.
- Check that we have new commands to add debud and slice tracks from
  a prompt.
- Check "Binder txn names + monitor contention on thread" context
  menu item still works.

Note: This CL also includes a bunch of changes to reduce the number of
circular imports as much as possible. One of them was causing the unit
tests to fail, but because it was difficult to say which one it was
a lot of them have just been removed.

Change-Id: If3b574395c39aab13ec795cb196600ef6fe0b7c4