metrics: Add metrics for Wattson app startup
Add metrics for Wattson for Android Performance to use on various app
startups tests. This also has a generic interface so that any future
metrics that are looking for a time period, similar to the app startup
metric which is specifically looking at the time period of the app
startup, can re-use most of the same metric definition.

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson.*'
Bug: 340938075
Change-Id: Ieb9c256e82d4fcd951b2b37b32a848704ff28aec
Signed-off-by: Samuel Wu <wusamuel@google.com>