ui: Use unique URIs for each "active cpu" track
Currently, these tracks have a single URI registered for the three
different track types (big, mid, little), and rely on track params in
the track state to instantiate the different versions of each track.

This patch just registers one track per CPU type (which is known up
front anyway).

This allows each track to be uniquely identified using its URI and
removes another dependency on track params in the track state which is
due to be removed.

Change-Id: I243b3c59f098cce5919a111e9b04c8de994a8052