Trace Redaction - Move construction into singleton
Previously redaction was configured in the main method. This meant it
was not possible to test the full pipeline. This was not an issue
because tests were focusing on individual primitives or small cluster
that were expected to work together.

This was not an issue when the pipeline was small, but now that the
pipeline is more complicated, there is a need to test the pipeline as
a whole - not just primitives.

This is the first step toward automating the adhoc manual tests.

Change-Id: Ibc829ccc50e7d427974c2d16cf73809124e24849