Trace Redaction - Move Redactor+Context out of fixture
The integration test fixture's core responsibility is to manage source
trace and redacted trace files. Some common values we add to it to
avoid declaring them in every test (redactor and context).

This coupling was okay when each test populated the redactor. However,
we'll be adding tests that use the production config and having this
silent default value has already caused issued when not reset.

To avoid this easy mistakes, this removes the redactor and context
from the fixture and requires each test to provide their own.

Change-Id: Ica0e0752bf2f092a914f0d3bc262242615ec0bd4