Trace Redaction - Reimplement and expand frame timeline tests
Because of the different (and similar) types of frame start events,
testing become very repetitive.

Before this code was implemented as a predicate, however when it
became the only code using the "runner", the two classes were merged
together. The tests were not updated.

This CL throws away the old tests and replaces them with a new suite
of tests that provider better test coverage - including the code that
modifies the packet.

Change-Id: I2745dd13a8af37099b9539fc62602c42800761e6