[stdlib]: Fix critical path correctness with kernel stalls
Kernel stalls during a critical path traversal could lead to
unexpected results. The expected behavior is that the critical
path should resume from the thread where the stall was first
encountered. This wasn't the case, the critical path would
naively follow the kernel stall until the entire critical path
interval regions has been resolved.

The new algorithm solves this in 3 phases:
1. Critical path following only userspace edges, i.e if a kernel
node is encountered, we jump to the prev_node instead of the waker_node.
This has the effect of staying on the thread when a kernel stall is
encountered.
2. Another critical path with the kernel nodes encountered from #1
as roots. This has the effect of filling the kernel stall gaps from
3. Flatten the intervals from the results of #1 and #2. Where any interval
in #2 is a child of the resulting interval from #1.

Fixed a small omission in the _intervals_to_roots macro where it wasn't
correctly bounded to a utid. It didn't affect correctness since we
post-filtered utid, but it affected performance.

Also removed a special case for IRQs. Prevoiusly, we skipped them because
we didn't have process attribution, this means they were not present in
the wakeup_graph and we couldn't easily answer questions like what part of
the critical path was stalled in an IRQ? Now, we don't skip them, we simply
follow their 'prev' edge, hence staying on the same thread, since we still
don't have process attribution. With this change, we have information about
nodes that were stalled waiting on an IRQ, even if we don't have any further
information about the IRQ.

Test: tools/diff_test_trace_processor.py out/android/trace_processor_shell --name-filter '.*thread_executing_span.*'

Change-Id: I0a451fbc218c2e1ae775eccd2e9966a69d1008a8