Add option to control print log
Two changes in this commit:
1. Add an option to control the print and priority in the logcat.
2. add stderr in the adb function to control the error output.

The default value are the same as before, so there is no logic changes.

Test:
run the script: ./gen_amalgamated_python_tools to generate the record_android_trace and tested with an Android device trace, all works as expected and all log print as before.


refactored the record android trace script

The main goal of this refactor is to improve readability and provide a more modular structure that can be customized and incorporated into other scripts. There is no logic change.

Test:
Change-Id: I7229ba3accdaad614440e43ba809db7cbf6bd001
run the script: ./gen_amalgamated_python_tools to generate the record_android_trace and tested with an Android device trace, all works as expected.