UI: set the window title on load
The logic to set the window title used to live
in sidebar.ts. This is because originally only the sidebar
needed to work out the title for itself, and later we added
some code to set the window title (by doing some hack that
avoids touching the dom too much by checking if the title
changed).
This CL just moves the logic to compute the title in
trace_controller and changes the title only in one place.

Bug: 348642298
Change-Id: I555e87deb59ade07e9980b0ad7281fae25c7e98f