ui: Improve sched slice details query efficiency
This patch replaces a slow LEFT JOIN with two queries, one of which
loads the data from the original table, and the other loads the data
from the joined table but using a simple INNER JOIN which is much
faster.

Testing:
Select a CPU slice in the CPU slice track and make sure we still have
working links to thread state slices in the details panel. It should be
much faster to load these details now as well.

Performance:

Before | After
--------------
~150ms | ~2ms

Change-Id: Iad0a2ebc837dec38b0b1cb7a0069efaa0804dfa7