Do not sort args explicitly in the UI.
This patch reverts aosp/2701235 while ensuring that
https://github.com/google/perfetto/issues/553 remains fixed.

The fundamental problem this is solving is ensuring that the order of
args is preserved, as some trace emitters might have an opinion on
the order of args.

The problem was that the args importer has been sorting the args to
ensure that each arg set is contigiuous (e.g. including args coming
from begin and end events of the same slice), but it was using the
interned key id to this.

This has been mostly working fine (as the strings are typically interned
in the correct order), but failed when some of the strings have already
been seen before.

Bug: b/348633950

Change-Id: Ie0d9c3be4dcde77d105b9bfd8351139861138ead