ui: Move COUNTER_TRACK_KIND from core_plugins -> core
COUNTER_TRACK_KIND is used by the core (i.e. controllers) as well as
from 'core' plugins. Currently it lives in the src/core_plugins dir
which means that the core depends on core_plugins which is something we
want to avoid.

This patch moves this constant into into the core.

Change-Id: I494555e2225c787bf1573ab9997129cda0fa5bed