ui: Associate Wattson estimate per thread
Associates Wattson estimate per thread based on the user-drawn UI
window. For all threads in the drawn UI window, all threads will be
grouped together across the x-axis (time) and y-axis (CPUs), and the
corresponding Wattson estimate will be summed together.

This feature allows users to attribute Wattson estimates on a thread
level, such that extraneous or irrelevant threads can be ignored OR
threads of interest can be selectively chosen for analysis.

Bug: 329507878
Change-Id: Iacd2aaf083ed7a88beaf6dc3c8f2427314a26981
Signed-off-by: Samuel Wu <wusamuel@google.com>